+++ README-Chinese.md
@@ -23,7 +23,7 @@ AS库目录有多个版本，如果希望查询不同版本的更新日志以及
     <dependency>
         <groupId>io.github.BeardedManZhao</groupId>
         <artifactId>algorithmStar</artifactId>
-        <version>1.28</version>
+        <version>1.29</version>
     </dependency>
 </dependencies>
 ```
