+++ src_code/src/main/java/zhao/algorithmMagic/MAIN1.java
@@ -1,4 +1,5 @@
 package zhao.algorithmMagic;
 
 public class MAIN1 {
+
 }
