+++ src_code/update/1.26_1.27.md
@@ -71,14 +71,14 @@ public class MAIN1 {
 package zhao.algorithmMagic;
 
 import zhao.algorithmMagic.core.AlgorithmStar;
-import zhao.algorithmMagic.core.MatixFactory;
+import zhao.algorithmMagic.core.MatrixFactory;
 import zhao.algorithmMagic.operands.matrix.DoubleMatrix;
 import zhao.algorithmMagic.operands.matrix.IntegerMatrix;
 
 public class MAIN1 {
     public static void main(String[] args) {
         // 使用 AS 门户类获取到矩阵工厂
-        final MatixFactory matixFactory = AlgorithmStar.matixFactory();
+        final MatrixFactory matixFactory = AlgorithmStar.matrixFactory();
         // 创建一个整形矩阵
         final IntegerMatrix integerMatrix = matixFactory.parseMatrix(
                 new int[]{1, 2, 3, 4},
@@ -112,7 +112,7 @@ public class MAIN1 {
         // 准备一张图的 URL
         final URL url = new URL("http://diskmirror.lingyuzhao.top/1/Binary/%E3%80%8AalgorithmStar%20%E6%9C%BA%E5%99%A8%E5%AD%A6%E4%B9%A0%E3%80%8B%E7%9A%84%E5%B0%81%E9%9D%A2.jpg");
         // 使用 AS 门户类将 URL 解析为 图片
-        final ColorMatrix colorMatrix = AlgorithmStar.matixFactory().parseImage(
+        final ColorMatrix colorMatrix = AlgorithmStar.matrixFactory().parseImage(
                 // 设置图片矩阵的类型
                 ColorMatrix.class,
                 // 设置构造操作中需要使用的参数
